Actually working ros2 message generation
