+++ roslibrust_test/tests/codegen_tests.rs
@@ -1,4 +1,4 @@
-use roslibrust_test::*;
+use roslibrust_test::ros1::*;
 
 /// Ensures that associate constants are generated on the test_msgs correctly
 /// requires test_msgs gen_code to have been generated.
